python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "symphony"
version = "0.1.0"
description = "Symphony turns project work into isolated, autonomous implementation runs, allowing teams to manage work instead of supervising coding agents."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastactor @ git+https://github.com/CyrusNuevoDia/fastactor@main",
  "codex-app-server-sdk",
  "httpx>=0.27",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "python-frontmatter>=1.1",
  "python-liquid>=1.12",
  "structlog>=24",
  "typer>=0.12",
]

[project.scripts]
symphony = "symphony.cli:app"

[dependency-groups]
dev = [
  "pytest>=8",
  "anyio[trio]>=4.13",
  "ruff",
  "pyright",
]

[tool.pytest.ini_options]
anyio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]

[tool.hatch.build.targets.wheel]
packages = ["src/symphony"]

[tool.hatch.metadata]
allow-direct-references = true
